Implement a process watcher, and partially plumb it
