+++ agent/conf.go
@@ -19,6 +19,7 @@ type Conf struct {
 
 type FileConf struct {
 	Path    string `toml:"path"`
+	Command string `toml:"command"`
 	Process string `toml:"process"`
 }
