+++ agent/mocks.go
@@ -15,6 +15,7 @@ func (m *MockClient) Heartbeat(_a0 string, _a1 Watchers) error {
 
 	return r0
 }
+
 func (m *MockClient) SendFile(_a0 string, _a1 string, _a2 []byte) error {
 	ret := m.Called()
 
@@ -22,6 +23,11 @@ func (m *MockClient) SendFile(_a0 string, _a1 string, _a2 []byte) error {
 
 	return r0
 }
+
+func (m *MockClient) SendProcessState(_a0 string, _a1 *watchedState) error {
+	return m.Called().Error(0)
+}
+
 func (m *MockClient) CreateServer(_a0 *Server) (string, error) {
 	return m.Called().String(0), nil
 }
