+++ test/data/spector.conf
@@ -1,0 +1,8 @@
+server_name = "spectortest"
+
+api_key = "XtG6yAdbsSJfMwgx9SwafhZ8"
+distro = "ubuntu"
+release = "16.04"
+
+[[files]]
+process = "dhclient"
